updates on PluginModal, making it more well integrated with discord's UI
