+++ src/components/PluginSettings/PluginModal.tsx
@@ -16,18 +16,13 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 
-import { generateId } from "@api/Commands";
 import { useSettings } from "@api/settings";
 import ErrorBoundary from "@components/ErrorBoundary";
 import { Flex } from "@components/Flex";
-import { LazyComponent } from "@utils/misc";
 import { ModalContent, ModalFooter, ModalHeader, ModalProps, ModalRoot, ModalSize } from "@utils/modal";
-import { proxyLazy } from "@utils/proxyLazy";
 import { OptionType, Plugin } from "@utils/types";
-import { findByCode, findByPropsLazy } from "@webpack";
-import { Button, FluxDispatcher, Forms, React, Text, Tooltip, UserStore, UserUtils } from "@webpack/common";
-import { User } from "discord-types/general";
-import { Constructor } from "type-fest";
+import { findByPropsLazy } from "@webpack";
+import { Button, Forms, React, Text, Tooltip } from "@webpack/common";
 
 import {
     ISettingElementProps,
@@ -39,27 +34,11 @@ import {
     SettingTextComponent
 } from "./components";
 
-const UserSummaryItem = LazyComponent(() => findByCode("defaultRenderUser", "showDefaultAvatarsForNullUsers"));
 const AvatarStyles = findByPropsLazy("moreUsers", "emptyUser", "avatarContainer", "clickableAvatar");
-const UserRecord: Constructor<Partial<User>> = proxyLazy(() => UserStore.getCurrentUser().constructor) as any;
 
 interface PluginModalProps extends ModalProps {
     plugin: Plugin;
     onRestartNeeded(): void;
-}
-
-/** To stop discord making unwanted requests... */
-function makeDummyUser(user: { name: string, id: BigInt; }) {
-    const newUser = new UserRecord({
-        username: user.name,
-        id: generateId(),
-        bot: true,
-    });
-    FluxDispatcher.dispatch({
-        type: "USER_UPDATE",
-        user: newUser,
-    });
-    return newUser;
 }
 
 const Components: Record<OptionType, React.ComponentType<ISettingElementProps<any>>> = {
@@ -73,8 +52,6 @@ const Components: Record<OptionType, React.ComponentType<ISettingElementProps<an
 };
 
 export default function PluginModal({ plugin, onRestartNeeded, onClose, transitionState }: PluginModalProps) {
-    const [authors, setAuthors] = React.useState<Partial<User>[]>([]);
-
     const pluginSettings = useSettings().plugins[plugin.name];
 
     const [tempSettings, setTempSettings] = React.useState<Record<string, any>>({});
@@ -83,17 +60,6 @@ export default function PluginModal({ plugin, onRestartNeeded, onClose, transiti
     const [saveError, setSaveError] = React.useState<string | null>(null);
 
     const canSubmit = () => Object.values(errors).every(e => !e);
-
-    React.useEffect(() => {
-        (async () => {
-            for (const user of plugin.authors.slice(0, 6)) {
-                const author = user.id
-                    ? await UserUtils.fetchUser(`${user.id}`).catch(() => makeDummyUser(user))
-                    : makeDummyUser(user);
-                setAuthors(a => [...a, author]);
-            }
-        })();
-    }, []);
 
     async function saveAndClose() {
         if (!plugin.options) {
@@ -150,48 +116,12 @@ export default function PluginModal({ plugin, onRestartNeeded, onClose, transiti
         return <Flex flexDirection="column" style={{ gap: 12 }}>{options}</Flex>;
     }
 
-    function renderMoreUsers(_label: string, count: number) {
-        const sliceCount = plugin.authors.length - count;
-        const sliceStart = plugin.authors.length - sliceCount;
-        const sliceEnd = sliceStart + plugin.authors.length - count;
-
-        return (
-            <Tooltip text={plugin.authors.slice(sliceStart, sliceEnd).map(u => u.name).join(", ")}>
-                {({ onMouseEnter, onMouseLeave }) => (
-                    <div
-                        className={AvatarStyles.moreUsers}
-                        onMouseEnter={onMouseEnter}
-                        onMouseLeave={onMouseLeave}
-                    >
-                        +{sliceCount}
-                    </div>
-                )}
-            </Tooltip>
-        );
-    }
-
     return (
-        <ModalRoot transitionState={transitionState} size={ModalSize.MEDIUM}>
-            <ModalHeader>
-                <Text variant="heading-md/bold">{plugin.name}</Text>
+        <ModalRoot transitionState={transitionState} size={ModalSize.DYNAMIC}>
+            <ModalHeader separator={false}>
+                <Text variant="heading-lg/semibold">{plugin.name}</Text>
             </ModalHeader>
-            <ModalContent style={{ marginBottom: 8, marginTop: 8 }}>
-                <Forms.FormSection>
-                    <Forms.FormTitle tag="h3">About {plugin.name}</Forms.FormTitle>
-                    <Forms.FormText>{plugin.description}</Forms.FormText>
-                    <div style={{ marginTop: 8, marginBottom: 8, width: "fit-content" }}>
-                        <UserSummaryItem
-                            users={authors}
-                            count={plugin.authors.length}
-                            guildId={undefined}
-                            renderIcon={false}
-                            max={6}
-                            showDefaultAvatarsForNullUsers
-                            showUserPopout
-                            renderMoreUsers={renderMoreUsers}
-                        />
-                    </div>
-                </Forms.FormSection>
+            <ModalContent style={{ marginBottom: 8, minWidth: "30vw", minHeight: "20vh" }}>
                 {!!plugin.settingsAboutComponent && (
                     <div style={{ marginBottom: 8 }}>
                         <Forms.FormSection>
@@ -202,7 +132,6 @@ export default function PluginModal({ plugin, onRestartNeeded, onClose, transiti
                     </div>
                 )}
                 <Forms.FormSection>
-                    <Forms.FormTitle tag="h3">Settings</Forms.FormTitle>
                     {renderSettings()}
                 </Forms.FormSection>
             </ModalContent>
